chore(SDK-417): e2e tests for Rust CDK (#246)
* SDK-417: e2e tests for Rust CDK

This change introduces a framework for writing end-to-end tests for
Rust CDK.

The tests use a simple test framework developed for the IC repo,
ic-state-machine-tests. The framework provides an easy-to-use
interface to messaging and execution layer. The Wasm execution happens
directly in the test process.

The compilation of ic-state-machine tests takes a while, but incremental
builds are quite fast (10 sec). A single test execution is also quite
fast (10 sec for debug test run, 2.5 sec in release test run).

The overall approach to testing:

  1. Write a simple canister that uses the desired feature and put it
     under tests/canisters.

  2. Add a test in tests/tests/e2e.rs that builds, installs, and
     interacts with the canister.

* sipmle -> simple

* add escargot 'print' feature

* add wasm32-unknown-unknown target

* Run e2e tests only on linux

* move e2e tests into a separate workflow

* update ic version and run e2e on macOS